# Conflicts:
#	aurorastation.dme

+++ code/game/objects/effects/chem/foam.dm
@@ -24,6 +24,12 @@
 	addtimer(CALLBACK(src, PROC_REF(tick)), 3 + metal * 3)
 	addtimer(CALLBACK(src, PROC_REF(post)), solid_time)
 
+	var/static/list/loc_connections = list(
+		COMSIG_ATOM_ENTERED = PROC_REF(on_entered),
+	)
+
+	AddElement(/datum/element/connect_loc, loc_connections)
+
 /obj/effect/effect/foam/proc/tick()
 	process()
 	checkReagents()
@@ -49,7 +55,7 @@
 	if(--amount < 0)
 		return
 
-	for(var/direction in GLOB.cardinal)
+	for(var/direction in GLOB.cardinals)
 		var/turf/T = get_step(src, direction)
 		if(!T)
 			continue
@@ -77,11 +83,13 @@
 
 		QDEL_IN(src, 5)
 
-/obj/effect/effect/foam/Crossed(var/atom/movable/AM)
+/obj/effect/effect/foam/proc/on_entered(datum/source, atom/movable/arrived, atom/old_loc, list/atom/old_locs)
+	SIGNAL_HANDLER
+
 	if(metal)
 		return
-	if(istype(AM, /mob/living))
-		var/mob/living/M = AM
+	if(istype(arrived, /mob/living))
+		var/mob/living/M = arrived
 		M.slip("the foam", 6)
 
 /obj/effect/effect/foam/spray
@@ -162,7 +170,11 @@
 /obj/structure/foamedmetal/ex_act(severity)
 	qdel(src)
 
-/obj/structure/foamedmetal/bullet_act()
+/obj/structure/foamedmetal/bullet_act(obj/projectile/hitting_projectile, def_zone, piercing_hit)
+	. = ..()
+	if(. != BULLET_ACT_HIT)
+		return .
+
 	if(metal == 1 || prob(50))
 		qdel(src)
 
@@ -238,4 +250,4 @@
 /obj/structure/foamedmetal/CanPass(atom/movable/mover, turf/target, height=1.5, air_group = 0)
 	if(air_group)
 		return 0
-	return !density
+	return ..()
